studymate/milestone3/requirements.txt
ibm-watsonx-ai>=1.1.8
python-dotenv>=1.0.1
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
pytest>=8.2.1
httpx>=0.27.0
pdfplumber>=0.11.0
# optional vector store libs
faiss-cpu>=1.8.0.post1



